% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_epic_vector.R
\name{close.forkable.meffilref}
\alias{close.forkable.meffilref}
\title{manage a meffilref}
\usage{
\method{close}{forkable.meffilref}(con, ...)
}
\arguments{
\item{con}{forkable.meffilref instance}

\item{\dots}{not used}
}
\description{
manage a meffilref
}
