% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{egage}
\alias{egage}
\title{subset of Horvath clock CpGs available on EPIC}
\format{
GRanges
}
\usage{
egage
}
\description{
subset of Horvath clock CpGs available on EPIC
}
\note{
derived from 13059_2013_3156_MOESM23_ESM.csv, supplemental to `https://genomebiology.biomedcentral.com/articles/10.1186/gb-2013-14-10-r115#Sec40`
}
\keyword{datasets}
