% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{epic_cpgs}
\alias{epic_cpgs}
\title{epic annotation of CpGs}
\format{
data.frame
}
\usage{
epic_cpgs
}
\description{
epic annotation of CpGs
}
\note{
Derived from IlluminaHumanMethylationEPICanno.ilm10b4.hg19.
}
\keyword{datasets}
