% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{epic_granges}
\alias{epic_granges}
\title{epic annotation of CpGs}
\format{
GRanges
}
\usage{
epic_granges
}
\description{
epic annotation of CpGs
}
\note{
Derived from IlluminaHumanMethylationEPICanno.ilm10b4.hg19.  Includes
in_prom in mcols, which gives the symbol for gene g when the associated CpG
is located in (TSS(g)-2000, TSS(g)+200)
}
\keyword{datasets}
