% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_epic_vector.R
\name{get_cpg}
\alias{get_cpg}
\title{get a named vector of betas from meffil GDS instance}
\usage{
get_cpg(x, cpg = "cg18478105", ncpu = 1)
}
\arguments{
\item{x}{instance of S3 class forkable.meffilref}

\item{cpg}{character(1) feature ID}

\item{ncpu}{integer(1) number of CPUs to use to pick through the GDS row data}
}
\description{
get a named vector of betas from meffil GDS instance
}
\note{
Row access is slow, so we use parallel ingestion of chunks.
}
\examples{
\dontrun{
tst1 = getmat()
tst1
head(get_cpg(tst1))
close(tst1)
}
}
