% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iccs.R
\name{get_iccs}
\alias{get_iccs}
\title{for a vector of indices into a SummarizedExperiment, extract ICCs for a GLS model}
\usage{
get_iccs(
  se,
  inds = 1:10,
  makedf.in = makedf_copdgene,
  fixedfmla = beta ~ race + gender + smkstat,
  ...
)
}
\arguments{
\item{se}{SummarizedExperiment}

\item{inds}{numeric() row indices}

\item{makedf.in}{function that generates a data.frame for use by nlme:gls, consumes se and dots,
must return a data.frame with beta and id as columns, fixedfmla will use beta and any
columns other than id to build model}

\item{fixedfmla}{formula with beta as dependent variable}

\item{\dots}{passed to makedf.in, within sapply}
}
\description{
for a vector of indices into a SummarizedExperiment, extract ICCs for a GLS model
}
\examples{
requireNamespace("HDF5Array")
requireNamespace("SummarizedExperiment")
se6 = HDF5Array::loadHDF5SummarizedExperiment(system.file("chr6_hdf5", package="meffilTools"))
makedf_simple = function(se, ...) {
  stopifnot(nrow(se)==1)
  data.frame(beta=as.numeric(SummarizedExperiment::assay(se)),
       id=as.numeric(factor(se$Sample_Group)))
}
gg = get_iccs(se6, inds=1:5, makedf.in=makedf_simple, fixedfmla=beta~1)
gg
}
