% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iccs_old.R
\name{get_iccs_old}
\alias{get_iccs_old}
\title{for a vector of indices into a SummarizedExperiment, extract ICCs for a GLS model}
\usage{
get_iccs_old(se, inds = 1:10, fixedfmla = beta ~ race + gender + smkstat)
}
\arguments{
\item{se}{SummarizedExperiment}

\item{inds}{numeric() row indices}

\item{fixedfmla}{formula with beta as dependent variable}
}
\description{
for a vector of indices into a SummarizedExperiment, extract ICCs for a GLS model
}
