% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{horvathCpGs}
\alias{horvathCpGs}
\title{353 CpGs identified in Horvath S Genome Biology 2013}
\format{
character vector
}
\usage{
horvathCpGs
}
\description{
353 CpGs identified in Horvath S Genome Biology 2013
}
\note{
not all are present on EPIC array
}
\keyword{datasets}
