% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iccs.R
\name{makedf_copdgene}
\alias{makedf_copdgene}
\title{make a data.frame tailored to COPDgene EPIC arrays, for fitting nlme::gls}
\usage{
makedf_copdgene(
  se,
  idvar = "SID",
  gendvar = "Gender",
  racevar = "race",
  phasevar = "Phase_study",
  smkvar = "smoking_status",
  agevar = "age_visit",
  sampname = "Sample.name"
)
}
\arguments{
\item{se}{SummarizedExperiment with one CpG}

\item{idvar}{character(1) ID in colData(se)}

\item{gendvar}{character(1)}

\item{racevar}{character(1)}

\item{phasevar}{character(1)}

\item{smkvar}{character(1)}

\item{agevar}{character(1)}

\item{sampname}{character(1)}
}
\description{
make a data.frame tailored to COPDgene EPIC arrays, for fitting nlme::gls
}
\note{
Designed to work with the HDF5SummarizedExperiments made from COPDGene EPIC arrays
}
