% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_epic_vector.R
\name{print.forkable.meffilref}
\alias{print.forkable.meffilref}
\title{print meth}
\usage{
\method{print}{forkable.meffilref}(x, ...)
}
\arguments{
\item{x}{instance of forkable.meffilref}

\item{\dots}{not used}
}
\description{
print meth
}
