% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_epic_vector.R
\name{pull_epic_vector}
\alias{pull_epic_vector}
\title{grab a vector of betas assuming epic dimensions}
\usage{
pull_epic_vector(
  colid = "TOE440084_204234730121_R06C01_379130",
  cpg = NULL,
  base = "/udd/resiq/meffil/QC/COPD.blood",
  gds = "beta.copd.blood.gds",
  ncpg = 865859L,
  nsamps = 12147L
)
}
\arguments{
\item{colid}{character(1) or NULL column id to select, assumes 'col.names' present}

\item{cpg}{character(1) or NULL CpG to select, assumes 'row.names' present}

\item{base}{character(1) folder in which .gds file is stored}

\item{gds}{character(1) local name of .gds file}

\item{ncpg}{integer(1) number of cpgs, the 'row' extent of 'matrix', assumed present in gds}

\item{nsamps}{integer(1) number of columns, the 'col' extent of 'matrix'}
}
\description{
grab a vector of betas assuming epic dimensions
}
\note{
This function uses openfn.gds and on.exit(closefn.gds(...)), extracts a
matrix and assigns dimnames.  We might want to make dimnames assignment optional.
}
\examples{
\dontrun{
x = pull_epic_vector()
stopifnot(nrow(x) == 865859)
head(x)
y = pull_epic_vector(cpg="cg18478105", colid=NULL)
head(y)
}
}
