% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{rept_filtse}
\alias{rept_filtse}
\title{create a report on a filtered SummarizedExperiment}
\usage{
rept_filtse(se, sampvar = "Sample_Group", bspar = IrlbaParam(), npc = 6)
}
\arguments{
\item{se}{SummarizedExperiment}

\item{sampvar}{character(1) name of a variable in colData to be regarded as a factor}

\item{bspar}{instance of BiocSingularParam, defaults to `IrlbaParam()`}

\item{npc}{numeric(1) number of PCs to be computed (passed to runPCA as `rank`)}
}
\description{
create a report on a filtered SummarizedExperiment
}
