% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_to_promoters.R
\name{subset_to_promoters}
\alias{subset_to_promoters}
\title{use epic_granges annotation to filter an EPIC RangedSummarizedExperiment to promoter regions
and add associated gene symbols}
\usage{
subset_to_promoters(se)
}
\arguments{
\item{se}{RangedSummarizedExperiment with EPIC data}
}
\description{
use epic_granges annotation to filter an EPIC RangedSummarizedExperiment to promoter regions
and add associated gene symbols
}
