% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{summary.reportObj}
\alias{summary.reportObj}
\title{summary for report}
\usage{
\method{summary}{reportObj}(object, ...)
}
\arguments{
\item{object}{instance of reportObj}

\item{\dots}{not used}
}
\description{
summary for report
}
